requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.0
easyocr==1.7.2
efficientnet-pytorch==0.7.1
filelock==3.16.1
flask==3.1.0
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
imageio==2.36.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
kiwisolver==1.4.7
lazy-loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
regex==2024.11.6
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sympy==1.13.1
tifffile==2024.8.30
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.32
ultralytics-thop==2.0.11
urllib3==2.2.3
werkzeug==3.1.3
zipp==3.21.0
